Committing a fix here as what I believe to be a better workaround for https://github.com/GradConnection/django-datetimezone-field/commit/62bf52eff11ac838ec18d548bf2c54c2b1a6191f
Particularly, that fix doesn't account for the use case of keeping the exact same time and date while changing the datetime. Instead of ignoring the time zone when comparing if the field changed, I changed the widget to check if both the date and time fields are empty, and if so, set the timezone to empty as well, as a timezone by itself (with neither a date nor a time) makes no sense.

If this is sensible, perhaps SplitTimeTimeZoneWidget needs a similar implementation, checking only the time field? What do you think?